chore(ssi): extract injection validation to testutils
This commit extracts the pod and container validation logic to a
seperate package so that it can be imported in our end to end tests.
